ODH-1835 Extend Stagis parser - Task info - (supposedly) fixed Pickle error
